Use uint16_t for UTF-16 code points in `utf16::{lead,trail}_surr`.
Previously uint32_t was used to avoid static_cast on 16-bit results of
arithmetic operations on 32-bit values.

+++ src/encoding/utf16.cc
@@ -108,15 +108,15 @@ static void split_by_rune_length(RangeSuffix*& root, IrAllocator& alc, rune l, r
             add_continuous1(root, alc, l, h);
         } else {
             add_continuous1(root, alc, l, MAX_1WORD_RUNE);
-            const uint32_t h_ld = lead_surr(h);
-            const uint32_t h_tr = trail_surr(h);
+            const uint16_t h_ld = lead_surr(h);
+            const uint16_t h_tr = trail_surr(h);
             split_by_continuity(root, alc, MIN_LEAD_SURR, h_ld, MIN_TRAIL_SURR, h_tr);
         }
     } else {
-        const uint32_t l_ld = lead_surr(l);
-        const uint32_t l_tr = trail_surr(l);
-        const uint32_t h_ld = lead_surr(h);
-        const uint32_t h_tr = trail_surr(h);
+        const uint16_t l_ld = lead_surr(l);
+        const uint16_t l_tr = trail_surr(l);
+        const uint16_t h_ld = lead_surr(h);
+        const uint16_t h_tr = trail_surr(h);
         split_by_continuity(root, alc, l_ld, h_ld, l_tr, h_tr);
     }
 }
